requirements.txt
africastalking==2.0.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.10.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.7
django-phonenumber-field==8.4.0
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.11
multidict==6.7.0
packaging==25.0
phonenumberslite==9.0.19
pillow==12.0.0
propcache==0.4.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
PyYAML==6.0.3
requests==2.32.5
responses==0.25.8
schema==0.7.8
sqlparse==0.5.3
twilio==9.8.7
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
yarl==1.22.0